change package name

+++ debug-install.sh
@@ -1,18 +1,20 @@
 #!/bin/bash
 WORK_DIR=$(cd `dirname $0`; pwd)
 cd $WORK_DIR
-VENV_NAME=$(crudini --get setup.cfg metadata conda_venv_name)
-python_path=$(which python)
-venv_path=$(dirname $(dirname $python_path))
 
-if [ "/root/miniconda3/envs/$VENV_NAME" = "$venv_path" ]; then
-  echo "in $VENV_NAME env, start work"
-else
-  echo "ERROR: not in $VENV_NAME env, exit"
-  exit 1
+# check if have crudini
+if ! command -v crudini &> /dev/null
+then
+    echo "crudini could not be found"
+    echo "please install crudini first"
+    exit
 fi
 
-bash make.sh
 package_base_name=$(crudini --get setup.cfg metadata name)
 pip uninstall $package_base_name -y
+
+rm -rf build
+rm -rf dist
+python3 setup.py sdist
+
 pip install dist/$package_base_name-*.tar.gz
